Move routes/api to src/routes
In an effort to reduce the complexity of the backend there is no need to have
routes separately (since we're not planning to have several routes).
